Replace incomingCache with objectStorage
