+++ plugins/gossip/packet_processor.go
@@ -12,15 +12,12 @@ import (
 
 	"github.com/iotaledger/hive.go/batchhasher"
 	"github.com/iotaledger/hive.go/daemon"
-	"github.com/iotaledger/hive.go/lru_cache"
 	"github.com/iotaledger/hive.go/math"
-	"github.com/iotaledger/hive.go/syncutils"
-	"github.com/iotaledger/hive.go/typeutils"
+	"github.com/iotaledger/hive.go/objectstorage"
 	"github.com/iotaledger/hive.go/workerpool"
 
 	"github.com/gohornet/hornet/packages/compressed"
 	"github.com/gohornet/hornet/packages/model/hornet"
-	"github.com/gohornet/hornet/packages/model/milestone_index"
 	"github.com/gohornet/hornet/packages/model/queue"
 	"github.com/gohornet/hornet/packages/model/tangle"
 	"github.com/gohornet/hornet/packages/profile"
@@ -36,15 +33,16 @@ var (
 	packetProcessorWorkerCount = batchhasher.CURLP81.GetBatchSize() * batchhasher.CURLP81.GetWorkerCount()
 	packetProcessorWorkerPool  *workerpool.WorkerPool
 
-	RequestQueue  *queue.RequestQueue
-	IncomingCache *lru_cache.LRUCache
+	RequestQueue *queue.RequestQueue
 
 	ErrTxExpired = errors.New("tx too old")
 )
 
 func configurePacketProcessor() {
+	opts := profile.GetProfile().Caches.IncomingTransactionFilter
+
 	RequestQueue = queue.NewRequestQueue()
-	IncomingCache = lru_cache.NewLRUCache(profile.GetProfile().Caches.IncomingTransactionFilter.Size)
+	incomingStorage = objectstorage.New(nil, incomingFactory, objectstorage.CacheTime(time.Duration(opts.CacheTimeMs)*time.Millisecond), objectstorage.PersistenceEnabled(false))
 
 	gossipLogger.Infof("Configuring packetProcessorWorkerPool with %d workers", packetProcessorWorkerCount)
 	packetProcessorWorkerPool = workerpool.New(func(task workerpool.Task) {
@@ -79,179 +77,6 @@ func runPacketProcessor() {
 		packetProcessorWorkerPool.StopAndWait()
 		gossipLogger.Info("Stopping PacketProcessor ... done")
 	}, shutdown.ShutdownPriorityPacketProcessor)
-}
-
-type NeighborRequest struct {
-	p                          *protocol
-	reqHashBytes               []byte
-	reqMilestoneIndex          milestone_index.MilestoneIndex
-	isMilestoneRequest         bool
-	isTransactionRequest       bool
-	isLegacyTransactionRequest bool
-	hasNoRequest               bool
-}
-
-func (n *NeighborRequest) punish() {
-	n.p.Neighbor.Metrics.IncrInvalidTransactionsCount()
-}
-
-func (n *NeighborRequest) notify(recHashBytes []byte) {
-	n.p.Neighbor.Reply(recHashBytes, n)
-}
-
-type PendingNeighborRequests struct {
-	startProcessingLock syncutils.Mutex
-
-	// data
-	dataLock     syncutils.RWMutex
-	recTxBytes   []byte
-	recHashBytes []byte
-	recHash      trinary.Hash
-	hornetTx     *hornet.Transaction
-
-	// status
-	statusLock syncutils.RWMutex
-	invalid    bool
-	hashing    bool
-
-	// requests
-	requestsLock syncutils.RWMutex
-	requests     []*NeighborRequest
-}
-
-func (p *PendingNeighborRequests) AddLegacyTxRequest(neighbor *protocol, reqHashBytes []byte) {
-	p.requestsLock.Lock()
-	defer p.requestsLock.Unlock()
-
-	p.requests = append(p.requests, &NeighborRequest{
-		p:                          neighbor,
-		reqHashBytes:               reqHashBytes,
-		isLegacyTransactionRequest: true,
-	})
-}
-
-func (p *PendingNeighborRequests) BlockFeedback(neighbor *protocol) {
-	p.requestsLock.Lock()
-	defer p.requestsLock.Unlock()
-
-	p.requests = append(p.requests, &NeighborRequest{
-		p:            neighbor,
-		hasNoRequest: true,
-	})
-}
-
-func (p *PendingNeighborRequests) IsHashing() bool {
-	p.statusLock.RLock()
-	defer p.statusLock.RUnlock()
-	return p.hashing
-}
-
-func (p *PendingNeighborRequests) IsHashed() bool {
-	p.statusLock.RLock()
-	defer p.statusLock.RUnlock()
-	return len(p.recHashBytes) > 0
-}
-
-func (p *PendingNeighborRequests) IsInvalid() bool {
-	p.statusLock.RLock()
-	defer p.statusLock.RUnlock()
-	return p.invalid
-}
-
-func (p *PendingNeighborRequests) GetTxHash() trinary.Hash {
-	p.dataLock.RLock()
-	defer p.dataLock.RUnlock()
-	return p.recHash
-}
-
-func (p *PendingNeighborRequests) GetTxHashBytes() []byte {
-	p.dataLock.RLock()
-	defer p.dataLock.RUnlock()
-	return p.recHashBytes
-}
-
-func (p *PendingNeighborRequests) process() {
-	p.startProcessingLock.Lock()
-
-	if p.IsHashing() {
-		p.startProcessingLock.Unlock()
-		return
-	} else if p.IsInvalid() {
-		p.startProcessingLock.Unlock()
-		p.punish()
-		return
-	} else if p.IsHashed() {
-		p.startProcessingLock.Unlock()
-
-		// Mark the pending request as received because we received the requested Tx Hash
-		requested := RequestQueue.MarkReceived(p.hornetTx.Tx.Hash)
-
-		if requested {
-			// Tx is requested => ignore that it was marked as stale before
-			p.hornetTx.SetRequested(requested)
-			Events.ReceivedTransaction.Trigger(p.hornetTx)
-		}
-
-		p.notify()
-		return
-	}
-
-	p.statusLock.Lock()
-	p.hashing = true
-	p.statusLock.Unlock()
-	p.startProcessingLock.Unlock()
-
-	tx, err := compressed.TransactionFromCompressedBytes(p.recTxBytes)
-	if err != nil {
-		return
-	}
-
-	if !transaction.HasValidNonce(tx, ownMWM) {
-		// PoW is invalid => punish neighbor
-		p.statusLock.Lock()
-		p.invalid = true
-		p.statusLock.Unlock()
-
-		// Do not answer
-		p.punish()
-		return
-	}
-
-	// Mark the pending request as received because we received the requested Tx Hash
-	requested := RequestQueue.MarkReceived(tx.Hash)
-
-	// POW valid => Process the message
-	hornetTx := hornet.NewTransactionFromGossip(tx, p.recTxBytes, requested)
-
-	// received tx was not requested and has an invalid timestamp (maybe before snapshot?)
-	// => do not store in our database
-	// => we need to reply to answer the neighbors request
-	timeValid, broadcast := checkTimestamp(hornetTx)
-	stale := !requested && !timeValid
-	recHashBytes := trinary.MustTrytesToBytes(tx.Hash)[:49]
-
-	p.dataLock.Lock()
-	p.recHash = tx.Hash
-	p.recHashBytes = recHashBytes
-	p.hornetTx = hornetTx
-	p.dataLock.Unlock()
-
-	p.statusLock.Lock()
-	p.hashing = false
-	p.statusLock.Unlock()
-
-	if !stale {
-		// Ignore stale transactions until they are requested
-		Events.ReceivedTransaction.Trigger(hornetTx)
-
-		if !requested && broadcast {
-			p.broadcast()
-		}
-	} else if len(p.requests) == 1 {
-		p.requests[0].p.Neighbor.Metrics.IncrInvalidTransactionsCount()
-	}
-
-	p.notify()
 }
 
 func BroadcastTransactionFromAPI(txTrytes trinary.Trytes) error {
@@ -302,60 +127,6 @@ func BroadcastTransactionFromAPI(txTrytes trinary.Trytes) error {
 	return nil
 }
 
-func (p *PendingNeighborRequests) notify() {
-	p.requestsLock.Lock()
-
-	for _, n := range p.requests {
-		n.notify(p.recHashBytes)
-	}
-
-	p.requests = make([]*NeighborRequest, 0)
-	p.requestsLock.Unlock()
-}
-
-func (p *PendingNeighborRequests) punish() {
-	p.requestsLock.Lock()
-
-	for _, n := range p.requests {
-		// Tx is known as invalid => punish neighbor
-		server.SharedServerMetrics.IncrInvalidTransactionsCount()
-		n.p.Neighbor.Metrics.IncrInvalidTransactionsCount()
-		n.punish()
-	}
-
-	p.requests = make([]*NeighborRequest, 0)
-	p.requestsLock.Unlock()
-}
-
-func (p *PendingNeighborRequests) broadcast() {
-	p.requestsLock.RLock()
-
-	excludedNeighbors := make(map[string]struct{})
-	for _, neighbor := range p.requests {
-		excludedNeighbors[neighbor.p.Neighbor.Identity] = struct{}{}
-	}
-	p.requestsLock.RUnlock()
-
-	BroadcastTransaction(excludedNeighbors, p.recTxBytes, p.GetTxHashBytes())
-}
-
-func pendingRequestFor(recTxBytes []byte) (result *PendingNeighborRequests) {
-
-	cacheKey := typeutils.BytesToString(recTxBytes)
-
-	if cacheResult := IncomingCache.ComputeIfAbsent(cacheKey, func() interface{} {
-		return &PendingNeighborRequests{
-			recTxBytes: recTxBytes,
-			requests:   make([]*NeighborRequest, 0),
-		}
-	}); !typeutils.IsInterfaceNil(cacheResult) {
-		result = cacheResult.(*PendingNeighborRequests)
-	}
-	return
-}
-
-var genesisTruncatedBytes = make([]byte, NON_SIG_TX_PART_BYTES_LENGTH)
-
 func ProcessReceivedMilestoneRequest(protocol *protocol, data []byte) {
 	server.SharedServerMetrics.IncrReceivedMilestoneRequestsCount()
 	protocol.Neighbor.Metrics.IncrReceivedMilestoneRequestsCount()
@@ -389,7 +160,9 @@ func ProcessReceivedLegacyTransactionGossipData(protocol *protocol, data []byte)
 	reqHashBytes := ExtractRequestedTxHash(data)
 
 	txData := data[:txDataLen]
-	pending = pendingRequestFor(txData)
+	cachedRequest := GetCachedPendingNeighborRequest(txData) // +1
+	defer cachedRequest.Release()
+	pending = cachedRequest.GetRequest()
 	pending.AddLegacyTxRequest(protocol, reqHashBytes)
 	pending.process()
 }
@@ -401,7 +174,9 @@ func ProcessReceivedTransactionGossipData(protocol *protocol, txData []byte) {
 	protocol.Neighbor.Metrics.IncrAllTransactionsCount()
 
 	var pending *PendingNeighborRequests
-	pending = pendingRequestFor(txData)
+	cachedRequest := GetCachedPendingNeighborRequest(txData)
+	defer cachedRequest.Release()
+	pending = cachedRequest.GetRequest()
 	pending.BlockFeedback(protocol)
 	pending.process()
